requirements.txt
anthropic>=0.25.0
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
scikit-learn>=1.3.0
numpy>=1.24.0
tqdm>=4.65.0
flask>=3.0.0
flask-cors>=4.0.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.0
